% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate_and_align_dataframes.R
\name{adjust_factor_levels_order}
\alias{adjust_factor_levels_order}
\title{Adjust factor levels in observed data to match synthetic data}
\usage{
adjust_factor_levels_order(observed_data, synthetic_data)
}
\arguments{
\item{observed_data}{Dataframe needing factor level adjustment}

\item{synthetic_data}{Reference dataframe for factor levels}
}
\value{
Dataframe with adjusted factor levels and prints adjustments made.
}
\description{
Adjust factor levels in observed data to match synthetic data
}
\examples{
\dontrun{
df1 <- data.frame(a = factor(c("A", "B", "C", "A"), levels = c("A", "B", "C")), b = 1:4)
df3 <- data.frame(a = factor(c("A", "B", "C", "D"), levels = c("A", "B", "C", "D")), b = 1:4)
adjust_factor_levels_order(df1, df3)
}
}
\keyword{internal}
