% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-encode_categorical.R
\name{apply_encoding_column}
\alias{apply_encoding_column}
\title{Apply Encoding to Categorical Column}
\usage{
apply_encoding_column(column, encoding)
}
\arguments{
\item{column}{A categorical vector (factor or character) that needs to be encoded.}

\item{encoding}{A dataframe or tibble containing the encoding map. The first column should be
\code{feature_level}, and the remaining columns are numeric encodings (e.g., from PCA or mean encoding).}
}
\value{
A tibble with one or more numeric columns representing the encoded values of \code{column}.
The order of the rows is preserved.
}
\description{
Applies a precomputed encoding to a categorical variable. This merges the original values
with their corresponding encoded numerical representations, based on a lookup table.
}
\details{
The function merges the original values of \code{column} with the corresponding encodings
from the table \code{encoding}, through a \code{left_join}. Then it removes the categorical value.

Called by: \code{\link{encode_categorical}}.
}
