% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-privacy_evaluation.R
\name{calculate_min_hamming_distance}
\alias{calculate_min_hamming_distance}
\title{Calculate Minimum Hamming Distance}
\usage{
calculate_min_hamming_distance(sds_train, train_data, holdout_data)
}
\arguments{
\item{sds_train}{List that contains the synthetic dataframe.}

\item{train_data}{Training data.}

\item{holdout_data}{Holdout data.}
}
\value{
A dataframe with the minimum Hamming distances to the training and holdout data for each synthetic observation.
}
\description{
Calculates the minimum Hamming distance between each synthetic observation and the observations in both the training and holdout datasets.
}
\keyword{internal}
