% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-privacy_evaluation.R
\name{calculate_privacy_metrics}
\alias{calculate_privacy_metrics}
\title{Calculate Privacy Metrics}
\usage{
calculate_privacy_metrics(DCR)
}
\arguments{
\item{DCR}{A dataframe containing distance comparisons between synthetic and observed datasets.}
}
\value{
A named vector containing various privacy metrics.
}
\description{
Calculates various privacy metrics based on the distance comparisons between synthetic and observed datasets.
}
\keyword{internal}
