% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save_session_info.R
\name{capture_session_info}
\alias{capture_session_info}
\title{Capture session information with a timestamp}
\usage{
capture_session_info()
}
\value{
Session information.
}
\description{
Capture session information with a timestamp
}
\keyword{internal}
