% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate_and_align_dataframes.R
\name{check_column_count}
\alias{check_column_count}
\title{Check if the number of columns matches between two dataframes}
\usage{
check_column_count(df1, df2)
}
\arguments{
\item{df1}{First dataframe}

\item{df2}{Second dataframe}
}
\value{
Returns TRUE if the number of columns matches, otherwise FALSE
}
\description{
Check if the number of columns matches between two dataframes
}
\keyword{internal}
