% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate_and_align_dataframes.R
\name{check_column_names}
\alias{check_column_names}
\title{Check if the column names match between two dataframes}
\usage{
check_column_names(df1, df2)
}
\arguments{
\item{df1}{First dataframe}

\item{df2}{Second dataframe}
}
\value{
Returns TRUE if the column names match, otherwise FALSE
}
\description{
Check if the column names match between two dataframes
}
\keyword{internal}
