% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-utility_evaluation.R
\name{check_columns}
\alias{check_columns}
\title{Check if columns are equal (number of columns, column names)}
\usage{
check_columns(syn_object, data_observed)
}
\arguments{
\item{syn_object}{The synthetic data object as synthesized by using synthpop, with class synds}

\item{data_observed}{The observed data from which the synthesized data was created, with class dataframe}
}
\value{
An error if the number of columns or column names are not similar between syn_object and data_observed.
}
\description{
Check if columns are equal (number of columns, column names)
}
\keyword{internal}
