% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate_and_align_dataframes.R
\name{check_data_types}
\alias{check_data_types}
\title{Check if the data types match between two dataframes}
\usage{
check_data_types(df1, df2)
}
\arguments{
\item{df1}{First dataframe}

\item{df2}{Second dataframe}
}
\value{
Returns TRUE if the data types match, otherwise FALSE
}
\description{
Check if the data types match between two dataframes
}
\keyword{internal}
