% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-utility_evaluation.R
\name{check_df_data_observed}
\alias{check_df_data_observed}
\title{Check if data_observed is a data frame}
\usage{
check_df_data_observed(data_observed)
}
\arguments{
\item{data_observed}{The observed data from which the synthesized data was created, with class dataframe}
}
\value{
An error if data_observed is not a data frame.
}
\description{
Check if data_observed is a data frame
}
\keyword{internal}
