% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate_and_align_dataframes.R
\name{check_factor_level_order}
\alias{check_factor_level_order}
\title{Check if the order of levels matches for factor variables}
\usage{
check_factor_level_order(df1, df2)
}
\arguments{
\item{df1}{First dataframe}

\item{df2}{Second dataframe}
}
\value{
Returns TRUE if the order of levels matches for factor variables, otherwise FALSE
}
\description{
Check if the order of levels matches for factor variables
}
\keyword{internal}
