% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save_syn_data.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check if the file doesn't already exist in the specified path}
\usage{
check_file_exists(path, file_name, overwrite)
}
\arguments{
\item{path}{A specified path where the results will be saved.}

\item{file_name}{A name for the CSV and Rds file.}

\item{overwrite}{Should an already existing file with a similar file_name and
path be overwritten? (default = FALSE)}
}
\value{
An error if the file_name already exists in the specified path.
}
\description{
Check if the file doesn't already exist in the specified path
}
\keyword{internal}
