% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-load_packages.R
\name{check_installed_packages}
\alias{check_installed_packages}
\title{check installed packages}
\usage{
check_installed_packages(pkg)
}
\arguments{
\item{pkg}{A character vector of package names.}
}
\value{
A character vector of packages that are not installed.
}
\description{
check installed packages
}
\keyword{internal}
