% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save_syn_data.R
\name{check_not_empty}
\alias{check_not_empty}
\title{Check if the syn_object is not empty}
\usage{
check_not_empty(syn_object)
}
\arguments{
\item{syn_object}{A synthetic data object, resulting from the use of the Synthpop package.}
}
\value{
A warning if the syn_object is empty.
}
\description{
Check if the syn_object is not empty
}
\keyword{internal}
