% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save_syn_data.R
\name{check_syn}
\alias{check_syn}
\title{Check if the syn_object contains $syn}
\usage{
check_syn(syn_object)
}
\arguments{
\item{syn_object}{A synthetic data object, resulting from the use of the Synthpop package.}
}
\value{
An error if the syn_object does not contain $syn.
}
\description{
Check if the syn_object contains $syn
}
\keyword{internal}
