% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-utility_evaluation.R
\name{check_synds_syn_object}
\alias{check_synds_syn_object}
\title{Check if syn_object is a synds object}
\usage{
check_synds_syn_object(syn_object)
}
\arguments{
\item{syn_object}{The synthetic data object as synthesized by using synthpop, with class synds}
}
\value{
An error if syn_object is not of class synds.
}
\description{
Check if syn_object is a synds object
}
\keyword{internal}
