% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataframe_comparer.R
\name{combine_comparison_results}
\alias{combine_comparison_results}
\title{Combine Comparison Results}
\usage{
combine_comparison_results(real_data_stats, syn_data_stats, level_agreement)
}
\arguments{
\item{real_data_stats}{A tibble containing statistics for the real data frame.}

\item{syn_data_stats}{A tibble containing statistics for the synthetic data frame.}

\item{level_agreement}{A vector describing the level agreement between the data frames.}
}
\value{
A tibble that combines the results of both data frames side by side.
}
\description{
This helper function combines the statistical results from two data frames
(real and synthetic) side by side into a single tibble.
}
\keyword{internal}
