% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataframe_comparer.R
\name{compare_dataframes}
\alias{compare_dataframes}
\title{Compare Two Data Frames or Analyze a Single Data Frame}
\usage{
compare_dataframes(real_data, syn_data = NULL)
}
\arguments{
\item{real_data}{The main data frame to be analyzed or compared.}

\item{syn_data}{An optional second data frame for comparison. If NULL, only \code{real_data} is analyzed.}
}
\value{
A tibble containing the analysis results for either one or both data frames.
}
\description{
This helper function either analyzes a single data frame or compares two data frames.
If two data frames are provided, it ensures they have matching column names and the same number of columns.
}
\keyword{internal}
