% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate_and_align_dataframes.R
\name{compare_factor_levels}
\alias{compare_factor_levels}
\title{Check if the number of levels matches for factor variables}
\usage{
compare_factor_levels(df1, df2)
}
\arguments{
\item{df1}{First dataframe}

\item{df2}{Second dataframe}
}
\value{
Returns TRUE if the number of levels matches for factor variables, otherwise FALSE
}
\description{
Check if the number of levels matches for factor variables
}
\keyword{internal}
