% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save_session_info.R
\name{create_log_file_path}
\alias{create_log_file_path}
\title{Generate a file name with a timestamp}
\usage{
create_log_file_path(file_name, path)
}
\arguments{
\item{file_name}{Specifying the file name of the text file where the session info will be stored.
The timestamp will be added to the file name.}

\item{path}{The path where the log file information will be stored.}
}
\value{
A path including file name with a timestamp.
}
\description{
Generate a file name with a timestamp
}
\keyword{internal}
