% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_comparer.R
\name{dataframe_comparer}
\alias{dataframe_comparer}
\title{Compare Data Frames}
\usage{
dataframe_comparer(real_data, syn_data = NULL)
}
\arguments{
\item{real_data}{A data frame to be analyzed.}

\item{syn_data}{An optional second data frame for comparison with \code{real_data}.
If not provided, the function only analyzes \code{real_data}.}
}
\value{
A data frame summarizing the comparison, including variable types, counts of NA values,
empty values, unique values, and basic statistics (min, median, mean, max) for numeric columns.
If \code{syn_data} is provided, the comparison will include both data frames.
}
\description{
This function compares two data frames in terms of column types, count of NA values,
count of empty values, and number of unique values. If only one data frame is provided,
it returns these metrics for that data frame alone.
}
\examples{
# Analyzing a single data frame:
real_data <- data.frame(a = factor(c("low", "medium", "high")),
b = factor(c("red", "green", "blue"), levels = c("red", "green", "blue")),
c = factor(c("low", "medium", "high")),
d = factor(c("yes", "no", "yes")),
e = c(1.0, 2.5, 3.1),  # Numeric variable
f = c("apple", "banana", "cherry"),  # String variable
g = c(NA, NA, 5),  # Column with many NAs
h = c("", "", "filled")  # Column with many empty values
)
dataframe_comparer(real_data)

# Comparing two data frames:
syn_data <- data.frame(
a = factor(c("low", "medium", "high")),  # Exact Match
b = factor(c("blue", "red", "green"),
levels = c("blue", "red", "green")),  # Level Match, Order Mismatch
c = factor(c("low", "medium", "very high")),  # Level Mismatch
d = c("yes", "no", "yes"),  # Non-Factor
e = c(1.0, 2.7, 3.0),  # Numeric variable with slight differences
f = c("apple", "banana", "citrus"),  # String variable with a difference
g = c(NA, 10, NA),  # Column with different NAs
h = c("", "empty", "")  # Column with different empty values
)
dataframe_comparer(real_data , syn_data)
}
