% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUOpop-package.R
\docType{package}
\name{DUOpop-package}
\alias{DUOpop}
\alias{DUOpop-package}
\title{DUOpop: Simplifies data synthesis}
\description{
Streamlines Data Synthesis and Includes Various Metrics to Measure the Utility and Privacy of Synthetic Data.
}
\keyword{internal}
