% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_categorical.R
\name{encode_categorical}
\alias{encode_categorical}
\title{Encode Categorical Variable}
\usage{
encode_categorical(x, xp, target, feature_name, logger)
}
\arguments{
\item{x}{the observed (original) data column for the variable to encode (features).}

\item{xp}{the corresponding column in the already synthesized data.}

\item{target}{the target variable to be synthesized; used for deriving encodings.}

\item{feature_name}{the name (string) of the variable being encoded; used to  name resulting encoded columns to avoid collisions.}

\item{logger}{A function to handle log messages. This function should take 2 arguments. See \link[syn.EncodedCart]{syn.EncodedCart}}
}
\value{
A named list with two data frames:
\item{\code{x}}{Encoded version of the observed data.}
\item{\code{xp}}{Encoded version of the synthesized data.}
}
\description{
This function is called by \code{\link{syn.EncodedCart}} to prepare categorical predictors before synthesis.
It applies an encoding algorithm to a categorical variable in both the observed (\code{x})
and already synthesized (\code{xp}) data. Depending on the characteristics of the variable
and the target, this function automatically chooses between:
\itemize{
\item No encoding (if the number of levels is small)
\item Mean encoding
\item PCA with all components
\item PCA with a reduced number of components
}
}
\details{
If the number of levels in \code{x} is 15 or fewer, no encoding is applied. Otherwise, encoding is chosen based on the type and number of levels in 'target'. Column names of the encoded variables are prefixed with \code{feature_name} to prevent naming collisions.
This function is intended to be defined before the synthesis, as it is called during.

Internally calls: \code{\link{get_encoding}}, \code{\link{apply_encoding_column}}.

Called by: \code{\link{syn.EncodedCart}} to prepare categorical predictors before synthesis.
}
\seealso{
\code{\link{syn.EncodedCart}}
}
