% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_and_save_metadata.R
\name{extract_names_and_types}
\alias{extract_names_and_types}
\title{extract names and types}
\usage{
extract_names_and_types(data, lang = "en")
}
\arguments{
\item{data}{The data to extract the names and types of.}
}
\value{
nothing
}
\description{
extract the names of the fields and their datatypes from 'data'
}
\keyword{internal}
