% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_and_save_metadata.R
\name{get_datatype_translation}
\alias{get_datatype_translation}
\title{datatype translation}
\usage{
get_datatype_translation(lang)
}
\arguments{
\item{lang}{either "en" for english or "nl" for Dutch}
}
\value{
vector with translation of datatypes
}
\description{
get a vector of language specific type names
}
\keyword{internal}
