% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-privacy_evaluation.R
\name{handle_na}
\alias{handle_na}
\title{Replace NA Values in the Dataset}
\usage{
handle_na(data, na_value)
}
\arguments{
\item{data}{The dataset in which NA values should be replaced.}

\item{na_value}{The value to replace NA values with.}
}
\value{
A dataframe where all NA values have been replaced by \code{na_value}.
}
\description{
Replaces all NA values in the dataset with a specified value.
}
\keyword{internal}
