% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-calculate_dcr.R
\name{harmonize_factor_levels}
\alias{harmonize_factor_levels}
\title{Harmonize Factor Levels between two Data Frames}
\usage{
harmonize_factor_levels(data.x, data.y)
}
\arguments{
\item{data.x}{The first data frame.}

\item{data.y}{The second data frame.}
}
\value{
A list with two elements:
\describe{
\item{data.x}{The first data frame, with harmonized factor levels.}
\item{data.y}{The second data frame, with harmonized factor levels.}
}
}
\description{
This function ensures that factor columns with matching names in
two data frames have the same set of levels. It updates the factors in both data frames
so that all levels present in either data frame are represented in both.
}
\details{
The function checks for columns that are factors in both \code{data.x} and \code{data.y}
and harmonizes their levels. If a factor level is present in one data frame but not
in the other, it is added to both. Non-factor columns are left unchanged.

Called by \code{\link{calculate_DCR}}.
}
\examples{
df1 <- data.frame(color = factor(c("red", "blue"), levels = c("red", "blue")))
df2 <- data.frame(color = factor(c("blue", "yellow"), levels = c("blue", "yellow")))
result <- harmonize_factor_levels(df1, df2)
levels(result$data.x$color)
levels(result$data.y$color)

}
