% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-load_packages.R
\name{install_missing_packages}
\alias{install_missing_packages}
\title{check missing packages}
\usage{
install_missing_packages(new_pkg)
}
\arguments{
\item{new_pkg}{A character vector of packages that are not installed.}
}
\value{
Packages specified in new_pkg are installed.
}
\description{
check missing packages
}
\keyword{internal}
