% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load and install packages if necessary}
\usage{
load_packages(...)
}
\arguments{
\item{...}{the packages you want to install}
}
\value{
loaded packages
}
\description{
Specified packages are loaded and installed if necessary.
}
\examples{
\dontrun{
load_packages("tidyverse")
load_packages("tidyverse", "dplyr")
}
}
