% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-load_packages.R
\name{load_packages_helper}
\alias{load_packages_helper}
\title{load packages}
\usage{
load_packages_helper(pkg)
}
\arguments{
\item{pkg}{A character vector of package names.}
}
\value{
A logical vector indicating whether each package was successfully loaded.
}
\description{
load packages
}
\keyword{internal}
