% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-encode_categorical.R
\name{mean_encode}
\alias{mean_encode}
\title{Mean Encoding}
\usage{
mean_encode(x, y)
}
\arguments{
\item{x}{a categorical vector (factor) representing the predictor variable.}

\item{y}{a numerical vector representing the target variable.}
}
\value{
A dataframe with two columns:
\itemize{
\item \code{feature_level}: the level of the categorical variable \code{x}
\item \code{encoding_value}: the mean of \code{y} for each level of \code{x}
}
}
\description{
Computes a mean encoding for a categorical variable based on the average value of a numerical target.
}
\details{
This encoding is useful when the target is numeric and you want to preserve
target-related structure in the categorical predictor.

Called by \code{\link{get_encoding}}.
}
