% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save_syn_data.R
\name{notify_save_success}
\alias{notify_save_success}
\title{Notify the user about successful save}
\usage{
notify_save_success(file_name, path)
}
\arguments{
\item{file_name}{A name for the CSV and Rds file.}

\item{path}{A specified path where the results will be saved.}
}
\value{
A notification that the data is saved successfully, including the specified
file_name and path.
}
\description{
Notify the user about successful save
}
\keyword{internal}
