% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-encode_categorical.R
\name{pca_encode}
\alias{pca_encode}
\title{PCA Encoding}
\usage{
pca_encode(x, y)
}
\arguments{
\item{x}{a categorical vector (factor) of the predictor to be encoded.}

\item{y}{the categorical target variable.}
}
\value{
A named list with:
\itemize{
\item \code{encoding}: A tibble with one row per level of \code{x}, and one column per principal component. The first column is the \code{feature_level}.
\item \code{sdev}: The standard deviations of the principal components (from \code{prcomp}).
}
}
\description{
Transforms a categorical variable into one or more numeric features using Principal Component Analysis (PCA),
based on how the levels of the categorical predictor co-occur with levels of the categorical target.

First, a count matrix is created showing how often each level of \code{x} (predictor/feature) co-occurs with each level
of \code{y} (target). PCA is then applied to this matrix, and the result is used as a numerical encoding.
}
\details{
Levels of \code{x} that appear with similar distributions across \code{y} levels will have similar encodings.
The variance of the co-occurence matrix is used to decide whether scaling is applied.

Called by \code{\link{get_encoding}}.
}
