% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-utility_evaluation.R
\name{plot_utility_data}
\alias{plot_utility_data}
\title{Create a utility plot}
\usage{
plot_utility_data(
  utility_data,
  colors = c("darkgreen", "lightgreen", "yellow", "orange", "red")
)
}
\arguments{
\item{utility_data}{List resulting from running the function run_utility_tests()}
}
\value{
A utility plot.
}
\description{
Create a utility plot
}
\keyword{internal}
