% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_defaults.R
\name{print_info}
\alias{print_info}
\title{print info}
\usage{
print_info(type, data)
}
\description{
prints info about the encoding and synthesis process
This can be things such as which encoding and parameters are used.
}
