% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_defaults.R
\name{print_info_progress}
\alias{print_info_progress}
\title{print info and progress}
\usage{
print_info_progress(type, data)
}
\description{
comination of \link[print_info]{print_info} and \link[print_progress]{print_progress}
}
