% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_defaults.R
\name{print_performance}
\alias{print_performance}
\title{print performance}
\usage{
print_performance(type, data)
}
\description{
prints messages about the performance of the synthesis process
This can be things such as the time (in minutes) it takes to synthesise a column,
or how effective PCA was for a feature-target combination.
}
