% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_defaults.R
\name{print_progress}
\alias{print_progress}
\title{print progress}
\usage{
print_progress(type, data)
}
\description{
prints messages about the progress of the synthesis process.
This shows which columns are being encoded and are synthesised.
Can be usefull for debugging.
}
