% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_evaluation.R
\name{privacy_evaluation}
\alias{privacy_evaluation}
\title{Privacy Evaluation Metric}
\usage{
privacy_evaluation(
  data_observed,
  strata = NULL,
  split_ratio = 0.5,
  na_value = -1,
  ...
)
}
\arguments{
\item{data_observed}{The real dataset.}

\item{strata}{(Optional) A column name or vector of column names for stratification.}

\item{split_ratio}{Ratio for train-test split, default is 0.5.}

\item{na_value}{Value to replace NA's with, default is -1.}

\item{...}{Additional arguments passed to the synthesis function (e.g., \code{minstratumsize} for \code{syn.strata}).}
}
\value{
Privacy metrics.
}
\description{
Computes privacy metrics for synthetic data.
The function can perform either simple or stratified data synthesis based on the provided parameters.
}
\examples{
library(ggplot2)
data_observed <- data.frame(head(diamonds, 2000))
# Without stratification
privacy_evaluation(data_observed = data_observed)
# With stratification
data_observed <- data.frame(head(diamonds, 4500))
privacy_evaluation(data_observed = data_observed, strata = "cut")
}
