% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-utility_evaluation.R
\name{run_utility_tests}
\alias{run_utility_tests}
\title{Run utility tests}
\usage{
run_utility_tests(
  syn_object,
  data_observed,
  utility_stat = "S_pMSE",
  ngroups = 25
)
}
\arguments{
\item{syn_object}{The synthetic data object as synthesized by using synthpop, with class synds}

\item{data_observed}{The observed data from which the synthesized data was created, with class dataframe}

\item{utility_stat}{Parameter in the function compare.synds {synthpop}}

\item{ngroups}{Parameter in the function utility.tables {synthpop}}
}
\value{
A list containing 1- and 2-dimensional utility results.
}
\description{
Run utility tests
}
\keyword{internal}
