% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save_syn_data.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save data as CSV}
\usage{
save_as_csv(data, path, file_name, overwrite)
}
\arguments{
\item{data}{A synthetic data object including $syn, resulting from the use of the Synthpop package.}

\item{path}{A specified path where the results will be saved.}

\item{file_name}{A name for the CSV file.}

\item{overwrite}{Should an already existing file with a similar file_name and
path be overwritten? (default = FALSE)}
}
\value{
A CSV file from the synthetic data.
}
\description{
Save data as CSV
}
\keyword{internal}
