% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save_syn_data.R
\name{save_as_rds}
\alias{save_as_rds}
\title{Save data as Rds}
\usage{
save_as_rds(object, path, file_name, overwrite)
}
\arguments{
\item{object}{A synthetic data object including $syn, resulting from the use of the Synthpop package.}

\item{path}{A specified path where the results will be saved.}

\item{file_name}{A name for the Rds file.}

\item{overwrite}{Should an already existing file with a similar file_name and
path be overwritten? (default = FALSE)}
}
\value{
A Rds file from the synthetic data.
}
\description{
Save data as Rds
}
\keyword{internal}
