% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-privacy_evaluation.R
\name{split_data}
\alias{split_data}
\title{Split Dataset into Training and Holdout Sets}
\usage{
split_data(data, split_ratio)
}
\arguments{
\item{data}{The dataset to be split.}

\item{split_ratio}{The ratio for splitting the dataset into train and holdout sets. Default is 0.5.}
}
\value{
A list containing two dataframes: \code{train} and \code{holdout}.
}
\description{
Splits the dataset into training and holdout sets based on the provided split ratio.
}
\keyword{internal}
