% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_categorical.R
\name{syn.EncodedCart}
\alias{syn.EncodedCart}
\title{Encoded CART Synthesis Method}
\usage{
syn.EncodedCart(
  y,
  x,
  xp,
  smoothing = "",
  proper = FALSE,
  minbucket = 5,
  cp = 1e-08,
  ...
)
}
\arguments{
\item{y}{an original data vector of length {\code{n}} representing the target.}

\item{x}{a matrix ({\code{n}} x {\code{p}}) of original features used for fitting. \code{n} is the number of observations in the original data and \code{p} is the number of predictors.}

\item{xp}{a matrix ({\code{k}} x {\code{p}}) of synthesized covariates (from previous synthesis steps). \code{k} is the number of already synthesized rows and \code{p} is the number of predictors.}

\item{smoothing}{smoothing method for numeric variable. See \code{\link[synthpop]{syn.smooth}}.}

\item{proper}{for proper synthesis ({\code{proper = TRUE}}) a CART model is fitted to a bootstrapped sample of the original data.}

\item{minbucket}{the minimum number of observations in any terminal node. See \code{\link[rpart]{rpart.control}} for details.}

\item{cp}{complexity parameter. Any split that does not decrease the overall lack of fit by a factor of cp is not attempted. Small values of {\code{cp}} will grow large trees. See \code{\link[rpart]{rpart.control}} for details.}

\item{ctree.logger}{function(type,data) to handle log messages. Data is structured as JSON.
The following standard log functions have already been defined:
\itemize{
\item \link[print_nothing]{print_nothing}
\item \link[print_targets]{print_targets}
\item \link[print_info]{print_info}
\item \link[print_performance]{print_performance}
\item \link[print_progress]{print_progress}
\item \link[print_info_progress]{print_info_progress}
\item \link[print_all]{print_all}
}

If none of these fits your needs, you can define your own:
\code{custom_logger <- function(type, data) { #custom logging logic}}
\code{s1 <- syn(ods, method = "EncodedCart", ctree.logger = custom_logger)}
\code{type} is the topic of the log message. This can be: "info", "progress", "performance".
\code{data} contains the log output, in JSON format.}
}
\value{
A vector with synthesized values of \code{y}, generated using an encoded CART tree.
}
\description{
Applies CART synthesis after encoding categorical predictors using PCA or mean encoding.
This function wraps \code{\link[synthpop]{syn.cart}} and includes preprocessing for categorical variables
with many levels.

This function can be used a synthesis method in the \code{\link[synthpop]{syn.cart()}} framework by specifying \code{method = "EncodedCart"}.

All parameters that are relevant to the synthpop implementation of CART should be prefixed with "cart.".
}
\details{
Categorical variables in \code{x} that have many levels (more than 15) are encoded using
the \code{\link{encode_categorical}} function. This function chooses the appropriate transformation - mean encoding or principal component analysis - based
on the structure of the predictors and the target.

These encodings are applied to both \code{x} and \code{xp} to ensure consistency during synthesis.

Internally calls: \code{\link{encode_categorical}}, \code{\link[synthpop]{syn.cart}}.
}
\examples{
# Example usage within the synthpop framework:
#s1 <- syn(ods, method = "EncodedCart")

s2 <- syn(data_for_synthesis
      ,maxfaclevels = Inf
      ,method = "EncodedCart"
      ,cart.minbucket = 10
      ,cart.cp = 0.001
      ,ctree.logger = print_all)

}
\seealso{
\code{\link{encode_categorical}}, \code{\link[synthpop]{syn.cart}}, \code{\link[synthpop]{syn}}
}
