% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate_and_align_dataframes.R
\name{sync_data_classes}
\alias{sync_data_classes}
\title{Synchronize column class types between two dataframes}
\usage{
sync_data_classes(data1, data2)
}
\arguments{
\item{data1}{A dataframe.}

\item{data2}{A dataframe with the same number of columns as \code{data1}.}
}
\value{
A modified dataframe where columns have the same class types as
their corresponding columns in \code{data1}.
}
\description{
Synchronize column classes between two dataframes to match the first.
}
\examples{
\dontrun{
# Voorbeelddataframes met een datumkolom
df1 <- data.frame(
 a = factor(1:3),
 b = 1:3,
 c = letters[1:3],
 d = as.Date(c("2022-01-01", "2022-01-02", "2022-01-03")))
df2 <- data.frame(
 a = as.character(1:3),
 b = c('1', '2', '3'),
 c = LETTERS[4:6],
 d = as.character(c("2022-01-01", "2022-01-02", "2022-01-03")))  # Datum als karakterstring
df2_adjusted <- sync_data_classes(df1, df2)
}
}
\keyword{internal}
