% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-utility_evaluation.R
\name{synchronize_na_levels}
\alias{synchronize_na_levels}
\title{Synchronize NA Levels and Handle Missing Values in Datasets}
\usage{
synchronize_na_levels(observed_data, synthetic_data)
}
\arguments{
\item{observed_data}{A data frame containing the observed data.}

\item{synthetic_data}{A data frame containing the synthetic data.}
}
\value{
A list containing the modified \code{observed_data} and \code{synthetic_data}.
\describe{
\item{observed_data}{The modified observed data frame.}
\item{synthetic_data}{The modified synthetic data frame.}
}
}
\description{
This function synchronizes factor levels between \code{observed_data} and \code{synthetic_data},
replaces NA values with -999 in numeric variables of \code{synthetic_data},
and converts logical variables to factors under specific conditions.
}
\details{
The function performs the following operations:
\enumerate{
\item Synchronizes factor levels for factor or character variables that contain NA values.
\item Replaces NA values with -999 in numeric variables of \code{synthetic_data} that contain NA values.
\item Converts logical variables to factors in both datasets if the variable is logical in
\code{observed_data}, contains NA in \code{synthetic_data}, and does not contain NA in \code{observed_data}.
}
}
\keyword{internal}
