% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_evaluation.R
\name{utility_evaluation}
\alias{utility_evaluation}
\title{Utility test}
\usage{
utility_evaluation(
  syn_object,
  data_observed,
  colors = c("darkgreen", "lightgreen", "yellow", "orange", "red"),
  ngroups = 25,
  utility_stat = "S_pMSE",
  autofix = FALSE,
  print.flag = TRUE
)
}
\arguments{
\item{syn_object}{The synthetic data object as synthesized by using synthpop, with class synds}

\item{data_observed}{The observed data from which the synthesized data was created, with class dataframe}

\item{colors}{Five colors that will be used used in the utility plot.}

\item{ngroups}{Parameter in the function utility.tables {synthpop}}

\item{utility_stat}{Parameter in the function compare.synds {synthpop}}

\item{autofix}{Logical; if TRUE, automatically aligns and validates the synthetic and observed dataframes.}

\item{print.flag}{Parameter in the functions utility.tables and compare.synds {synthpop}}
}
\value{
A list of 1- and 2-dimensional utility visuals
}
\description{
This function measures the utility of a synthetic dataset by generating
1- and 2-dimensional (visual) assessments, using the SpMSE
}
\examples{
# Generate synthetic data and measure the utility
syn_object <- list(syn = data.frame(a = 1:3, b = 4:6), m = 1)
syn_object$syn$a <- as.factor(syn_object$syn$a)
syn_object$syn$b <- as.factor(syn_object$syn$b)
class(syn_object) <- "synds"
data_observed <- data.frame(a = 1:3, b = 4:6)
data_observed$a <- as.factor(data_observed$a)
data_observed$b <- as.factor(data_observed$b)
utility_evaluation(syn_object, data_observed)

# Example usage of utility_evaluation with mismatched factor levels
observed_data <- data.frame(
  a = factor(c("A", "B", "C"), levels = c("A", "B", "C")),
  b = 1:3
)
synthetic_data <- data.frame(
  a = factor(c("A", "B", "C"), levels = c("A", "C", "B")),
  b = c("3", "1", "2")
)
syn_object <- structure(list(syn = synthetic_data, m = 1), class = "synds")
# Align factor levels using autofix
result <- utility_evaluation(syn_object, observed_data, autofix = TRUE)
# Align factor levels using autofix
result <- utility_evaluation(syn_object, observed_data, autofix = TRUE)
}
