% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_and_align_dataframes.R
\name{validate_and_align_dataframes}
\alias{validate_and_align_dataframes}
\title{Validate and Align Two Data Frames}
\usage{
validate_and_align_dataframes(observed_data, synthetic_data)
}
\arguments{
\item{observed_data}{The observed data frame.}

\item{synthetic_data}{The synthetic data frame.}
}
\value{
An adjusted version of \code{synthetic_data}.
}
\description{
Validates and aligns two data frames by checking various properties.
}
\examples{
\dontrun{
observed_data <- data.frame(
  a = factor(c("A", "B", "C"), levels = c("A", "B", "C")),
  b = 1:3
)

synthetic_data <- data.frame(
  a = factor(c("A", "B", "C"), levels = c("A", "C", "B")),
  b = c("3", "1", "2")
)

synthetic_data_adjusted <- validate_and_align_dataframes(observed_data, synthetic_data)
}

}
