% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-privacy_evaluation.R
\name{validate_data}
\alias{validate_data}
\title{Validate the Dataset}
\usage{
validate_data(data)
}
\arguments{
\item{data}{The dataset to validate.}
}
\value{
Throws an error or warning if validation fails, otherwise returns NULL.
}
\description{
Validates the dataset by checking if it meets certain criteria.
Issues a warning if the dataset contains fewer than 1000 observations.
Stops execution if the dataset has zero observations.
}
\keyword{internal}
