% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save_session_info.R
\name{validate_path}
\alias{validate_path}
\title{Check if specified path exists}
\usage{
validate_path(path)
}
\arguments{
\item{path}{The path where the log file information will be stored.}
}
\value{
An error if the specified path does not exist.
}
\description{
Check if specified path exists
}
\keyword{internal}
