% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-save_syn_data.R
\name{validate_syn_object}
\alias{validate_syn_object}
\title{Validate the synthetic object}
\usage{
validate_syn_object(syn_object, path, file_name, overwrite)
}
\arguments{
\item{syn_object}{A synthetic data object, resulting from the use of the Synthpop package.}

\item{path}{A specified path where the results will be saved.}

\item{file_name}{A name for the CSV and Rds file.}

\item{overwrite}{Should an already existing file with a similar file_name and
path be overwritten? (default = FALSE)}
}
\value{
Warning if the syn_object is empty, and errors if the syn_object does not
contain $syn and if the specified file_name already exists in the specified path.
}
\description{
Validate the synthetic object
}
\keyword{internal}
