% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_and_save_metadata.R
\name{write_to_excel}
\alias{write_to_excel}
\title{write to excel}
\usage{
write_to_excel(data, path)
}
\value{
nothing
}
\description{
wrapper of write_xlsx to enable unit testing
}
\keyword{internal}
