% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create_and_save_metadata.R
\name{get_datatype_translation}
\alias{get_datatype_translation}
\title{Data Type Translation}
\usage{
get_datatype_translation(lang)
}
\arguments{
\item{lang}{Language in which the data types need to be saved. Either "en" for English or "nl" for Dutch.}
}
\value{
Vector with translation of data types.
}
\description{
Creates a vector of language specific data type names.
}
\keyword{internal}
